+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+## Added
+- Mempool detection support: see
+  `CompactBlockProcessor.{watchMempool,consumeMempoolStream,resolveMempools}`.
+
 # 2.3.2 - 2025-09-03
 
 ## Fixed